+++ value_set.go
@@ -1,6 +1,6 @@
 package puzzle
 
-func NewValueSet(valueCount uint8) (vs []bool) {
+func NewValueSet(valueCount uint) (vs []bool) {
 	vs = make([]bool, valueCount)
 	for i := range vs {
 		vs[i] = true
